+++ src/generation/UniversalPosition.cpp
@@ -4,8 +4,8 @@
 
 #include "generation/UniversalPosition.hpp"
 
-UniversalPosition::UniversalPosition(uint32_t seedUniverse, glm::vec3 galaxy, glm::vec3 solarSystem) : seedUniverse(seedUniverse), positionGalaxy(galaxy), positionSolarSystem(solarSystem) {}
-UniversalPosition::UniversalPosition(uint32_t seedUniverse, glm::vec3 galaxy) : seedUniverse(seedUniverse), positionGalaxy(galaxy), positionSolarSystem() {}
+UniversalPosition::UniversalPosition(uint32_t seedUniverse, glm::ivec3 galaxy, glm::ivec3 solarSystem) : seedUniverse(seedUniverse), positionGalaxy(galaxy), positionSolarSystem(solarSystem) {}
+UniversalPosition::UniversalPosition(uint32_t seedUniverse, glm::ivec3 galaxy) : seedUniverse(seedUniverse), positionGalaxy(galaxy), positionSolarSystem() {}
 UniversalPosition::UniversalPosition(uint32_t seedUniverse) : seedUniverse(seedUniverse), positionGalaxy(), positionSolarSystem() {}
 UniversalPosition::UniversalPosition() : seedUniverse(), positionGalaxy(), positionSolarSystem() {}
 
@@ -13,10 +13,10 @@ uint32_t UniversalPosition::getSeedUniverse() {
     return seedUniverse;
 }
 
-glm::vec3 UniversalPosition::getPositionGalaxy() {
+glm::ivec3 UniversalPosition::getPositionGalaxy() {
     return positionGalaxy;
 }
 
-glm::vec3 UniversalPosition::getPositionSolarSystem() {
+glm::ivec3 UniversalPosition::getPositionSolarSystem() {
     return positionSolarSystem;
 }
